pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coilbench"
version = "0.1.0"
description = "Benchmark suite for stellarator coil optimization algorithms."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "Alan Kaptanoglu", email = "alanakaptanoglu@gmail.com"}
]

dependencies = [
  "numpy>=1.24",
  "typer[all]>=0.9",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "black",
  "ruff",
]

[project.scripts]
coilbench = "coilbench.cli:main"

